Added Daylight Savings Time Functionality + bugfix
The cleaner.py code now has the ability to differentiate between when to
add daylight savings time to the UTC stamp.  A throwaway days column is
created that calculates elapsed days in fractions of hours since Jan 1
at midnight, and this is used to slice out the DST tweets where only a
four hour offset is needed.  This required a number of other bugfixes:
* 12 AM is set to 00 AM
* When indexing day for rollback past midnight, the offset was wrong
The last bug fix is important, because the index offset should have
been -2 to align to the proper point in the day and year arrays.

Other bugfixes:
* hashtag.py now has the hashtag and file name in the cspc.ini file.
* plot_hashtag.py now has correct year on hashtag.